Returns empty eyer object

EyerObject is a list with predetermined fields for consistent functioning of other functions within the Eyer package. Lots of functions in the eyer package can work on passed "raw" data, such as fixation dataframes etc., but the benefit of preprocessing and parsing data inside the eyer object is that the object is inherently validated and processed and functions know what to expect. This data science design has been implemented from similar approaches found in Matlab and Python packages.

EyerObject has the following fields

data

events

fixations

diameter

gaze

info: list with eyetracker specific settings, suhc as recording frequency, recorded eye etc.

start_time: start time of the first recording in datetime

resolution: monitor resolution as a list with width, height fields

eyetracker: string with name of the recording device

Fixations have obligatory columns: time (commonly time since start), x, y, duration Gaze have obligatory columns: time (commonly time since start), x, y Events have columns: time, x, y

EyerObject()

Value

eyer object